Revert "RayTracing"
This reverts commit c9c6e5ca45bad995dca8b4b18bd49117e1bcf638.

+++ GameTest/src/edu/wkashti/game/Calculate.java
@@ -102,7 +102,7 @@ public class Calculate extends Main {
 		if (i == iterations) {
 			return 0;
 		} else {
-			return (int) (i*255.0/iterations);
+			return i*255/iterations;
 		}
 	}
